Make mzn2fzn look for local includes in the correct location.
Multi-pass constructs a dummy model to simplify the task of switching
libraries during compilation. The includes in this dummy model are
included relative to the parent directory of the dummy model, which,
before this fix was the current working directory. This was incorrect
for models that included local (typically .model) files.
This fix sets the parent directory of this dummy model so that it
matches the original model to be compiled.