Fix the coercion of boolean sum in aggregation
The existence of a sum builtin for booleans with a body was causing
problems with the aggregation. The aggregation will replace sum calls
without evaluating the body. This meant that the bool2int coercion was
not performed. By removing the explicit versions the typechecker will
automatically add the coercions directly on the arguments.