Refactoring of MIPSolverinstance::processSearchAnnotation.
The modifications improve clarity of intent by flattening
the nested if conditions in the for loop, and additionally
giving variables better names now that more space is available.

The purpose is to open the function up for further modification
and extensions in case one of the solver backends needs more
sophisticated forms of search annotations from this function.

Refactored MIPSolverinstance::processSearchAnnotations to make intent more clear

Renamed and refactored processSearchAnnotations for MIP