Add cmake variable MZN_STATIC_STDLIB_DIR
This can be set when running cmake, so that MiniZinc will try to find its standard library in the given directory. If that fails (or is not set), it will continue the search starting from the path of the binary, as before. The static directory can still be overridden by command line options or the MZN_STDLIB_DIR environment variable.
Fixes #421.