Fix bug that prevented type-checking of type-inst expressions in let variables.
The bug was caused by determining the type of all type-insts of all variables too early. Let-bound variables inside generators were checked before the type of the generator was determined. This commit fixes the order in which the expressions in a let are type-checked, and makes sure that the type of a variable declaration is determined only after all types it depends on are known.